Simplify data model
I can't keep trying to figure out what the data model should be.
I can't know until we write front-end code to display it, and we know
what we want to display.

So - until then I'm just dropping all the data into an array. we'll
query it by hand, do all the calculations on the server, and serve it up
slow.
Then we'll go back around and align mongo with what we want to do!