EPIC COMMIT! Auth and Vanity Pages
![My Reaction to this Commit](http://i.imgur.com/F9AEB.gif)

Added authentication via devise. Created a user model. Now you have to
be authenticated if you want to use the user_application controller at
all.

Wanted to have unauthenticated pages to look at as well, so the easiest
thing was to make a new "Static" controller and dropped in a couple of
views there--the home and about pages.